* UI toolbar with config and exit buttons (only exit is functional) * fix several issues so that return-to-hbmenu works * fix restarting games again * fix 256-color palette update detection (wouldn't detect if all the entries were modified at once)

+++ source/ppu.h
@@ -166,7 +166,7 @@ typedef struct
 	u16 CGRAM[256];		// SNES CGRAM, xBGR1555
 	u16 Palette[256];	// our own palette, converted to RGBx5551
 	u8 PaletteUpdateCount[64];
-	u8 PaletteUpdateCount256;
+	u16 PaletteUpdateCount256;
 
 	u16 VRAMAddr;
 	u16 VRAMPref;
